Improve module file reading and writing
Use `flock()` to lock `.mod` files during reading and writing to prevent
writing while a read is in progress and reading or writing when a write
is in progress. To achieve this, change module file writing to use C
stdio so that the file descriptor is readily available. `errno` seems
to be set more reliably as well.

Change module file reading to only read the file once; we used to
read it to verify the checksum and then again to parse it.
Instead, change `Parsing` so that we can get the file contents
after `Prescan()` and use that to verify the checksum. Also, it has
a mechanism for searching directories for files, so make use of that
instead of duplicating that functionality in `ModFileReader`.
This requires some changes to how errors are returned so they can
be reported in the right place. The "isModuleFile" flag is passed in
to `SourceFile` so that file locking can be done only for module files.